Massive changes
- Refactor SettingsView into multiple files
- Refactor SettingsView+Help into QuickHelpView/
- Move internal settings to DeveloperView+Internal
- Add Quick Help entries for Compatibility and Developer
