+++ Malachite/Views/MalachiteAboutView.swift
@@ -75,21 +75,7 @@ struct MalachiteAboutView: View {
         .navigationTitle("view.title.about")
         .toolbar(content: {
             ToolbarItemGroup(placement: .topBarTrailing) {
-                if #available(iOS 26.0, *) {
-                    Button {
-                        self.dismissAction()
-                    } label: {
-                        Image(systemName: "checkmark")
-                            .tint(.primary)
-                    }
-                    .buttonStyle(.glassProminent)
-                } else {
-                    Button {
-                        self.dismissAction()
-                    } label: {
-                        Image(systemName: "checkmark.circle")
-                    }
-                }
+                MalachiteToolbarUtils(action: self.dismissAction, image: "checkmark", primary: true)
             }
         })
     }
