+++ Malachite/Views/SettingsView/SettingsView+About.swift
@@ -1,0 +1,59 @@
+//
+//  SettingsView+About.swift
+//  Malachite
+//
+//  Created by Eva Isabella Luna on 8/16/25.
+//
+
+import SwiftUI
+
+extension SettingsView {
+    struct About: View {
+        /// A State variable used for determining whether or not this view is being presented as a modal.
+        var dismissAction: (() -> Void)
+        /// A variable to hold the existing instance of ``MalachiteClassesObject``.
+        var utilities = MalachiteClassesObject()
+        
+        init(
+            utilities: MalachiteClassesObject,
+            dismissAction: @escaping (() -> Void)
+        ) {
+            self.utilities = utilities
+            self.dismissAction = dismissAction
+        }
+        
+        var body: some View {
+            Section {
+                MalachiteCellViewUtils(
+                    icon: "info.circle",
+                    disabled: nil,
+                    dangerous: false)
+                {
+                    NavigationLink(destination: MalachiteAboutView(dismissAction: dismissAction)) {
+                        Text("view.title.about")
+                    }
+                }
+                if utilities.versionType == "INTERNAL" {
+                    MalachiteCellViewUtils(
+                        icon: "checkmark.seal",
+                        disabled: nil,
+                        dangerous: false)
+                    {
+                        NavigationLink(destination: MalachiteCompatibilityView(dismissAction: dismissAction, utilities: utilities)) {
+                            Text("view.title.compatibility")
+                        }
+                    }
+                    MalachiteCellViewUtils(
+                        icon: "wrench.and.screwdriver",
+                        disabled: nil,
+                        dangerous: false)
+                    {
+                        NavigationLink(destination: DeveloperView(dismissAction: dismissAction)) {
+                            Text("view.title.developer")
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
